fixed vtk output
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@688 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/io/vtk/VTKMoleculeWriterImplementation.cpp
@@ -128,7 +128,7 @@ void VTKMoleculeWriterImplementation::initializeParallelVTKFile(const std::vecto
 	p_cells.PDataArray().push_back(p_cells_data);
 
 	PUnstructuredGrid_t p_unstructuredGrid(p_pointData, p_cellData, p_points, p_cells);
-	for (int i = 0; i < fileNames.size(); i++) {
+	for (size_t i = 0; i < fileNames.size(); i++) {
 		Piece p_piece(fileNames[i]);
 		p_unstructuredGrid.Piece().push_back(p_piece);
 	}
